test(acceptance): LineSectionAcceptanceTest 리팩터링
LineSectionAcceptanceTest 리팩터링
- LineSectionAcceptance 파일에 CRUD 함수 분리
- 통합 테스트 추가
